+++ src/Akeneo/Pim/Enrichment/Bundle/Storage/Sql/Completeness/SqlGetCompletenessProductMasks.php
@@ -50,7 +50,14 @@ final class SqlGetCompletenessProductMasks implements GetCompletenessProductMask
     public function fromProductIdentifiers(array $productIdentifiers): array
     {
         // TODO - TIP-1212: Replace the first LEFT JOIN (to pim_catalog_family) by an INNER JOIN
+        // PIM-9783: the initial query didn't use the CTE, we filtered directly the product in the main SELECT. There
+        // was some performance issues with a big number of productIdentifier. The CTE allows to fix it (please check
+        // the issue for further information).
         $sql = <<<SQL
+WITH
+filtered_product AS (
+    SELECT id FROM pim_catalog_product WHERE identifier IN (:productIdentifiers)
+)
 SELECT
     product.id AS id,
     product.identifier AS identifier,
@@ -60,11 +67,11 @@ SELECT
            COALESCE(pm2.raw_values, '{}'),
            product.raw_values
     ) AS rawValues
-FROM pim_catalog_product product
+FROM filtered_product
+    INNER JOIN pim_catalog_product product ON filtered_product.id = product.id
     LEFT JOIN pim_catalog_family family ON product.family_id = family.id
     LEFT JOIN pim_catalog_product_model pm1 ON product.product_model_id = pm1.id
     LEFT JOIN pim_catalog_product_model pm2 ON pm1.parent_id = pm2.id
-WHERE product.identifier IN (:productIdentifiers)
 GROUP BY product.identifier
 SQL;
 
@@ -110,9 +117,12 @@ SQL;
     {
         $attributeCodes = [];
         foreach ($rows as $row) {
-            $attributeCodes = array_merge($attributeCodes, array_keys($row['cleanedRawValues']));
+            // array_unique is important for big catalog (see PIM-9783), because for a very high number of rows the array_merge takes too much time.
+            // For instance for 1.000 rows it can take 1 second, for 10.000 rows more than 1 minute.
+            // With array_unique it's less than 1 second in both cases.
+            $attributeCodes = array_unique(array_merge($attributeCodes, array_keys($row['cleanedRawValues'])));
         }
-        $attributes = $this->getAttributes->forCodes(array_unique($attributeCodes));
+        $attributes = $this->getAttributes->forCodes($attributeCodes);
 
         $result = [];
         foreach ($rows as $row) {