+++ CHANGELOG.md
@@ -44,8 +44,10 @@
 - PIM-9765: Fix missing translation key in bulk actions when adding attributes values for some product
 - PIM-9780: Fix completed import/export job notification broken link
 - PIM-9783: Optimize batch query when compute completeness
+- PIM-9783: Optimize SQL query when compute completeness
 - PIM-9715: Prevent the deletion of an attribute used as a label by a family
 - PIM-9781: Fix Category tree not refreshing when switching locale
+- PIM-9779: Fix ACE order when loading ACLs
 - PIM-9739: Fix connection users, users, channels having a link to a sub-category
 
 ## New features